stdlib/wcsto*: Use wchar_t for wchars instead of int.
Then use casts to int once the value has been converted from wchar_t
to a small integer.

+++ newlib/libc/stdlib/wcstoimax.c
@@ -109,9 +109,9 @@ wcstoimax_l(const wchar_t * __restrict nptr,
 			c -= L'a' - 10;
 		else
 			break;
-		if (c >= base)
+		if ((int) c >= base)
 			break;
-		if (any < 0 || acc > cutoff || (acc == cutoff && c > cutlim))
+		if (any < 0 || acc > cutoff || (acc == cutoff && (int) c > cutlim))
 			any = -1;
 		else {
 			any = 1;
